Split the pipeline timeout tests and shield fail-fast teardown
The review round this responds to was filed against `2326b5a`, one commit
before the fix that addressed it; verification against head found all eight of
its findings already resolved. Two adjacent items were genuinely open, plus one
new CodeScene finding against the current head.

CodeScene flagged `test_pipeline` at 634 lines of code against a 600 limit,
with a Low Cohesion warning — caused by the timeout work landing in a module
otherwise about composition and streaming. Move the timeout-boundary and
teardown-cancellation tests into `test_pipeline_timeouts`, which addresses both
signals: `cs delta` reports 8.03 -> 10.00 with "Lines of Code in a Single File"
fixed, and the new module scores 10.00 too.

Shield the fail-fast teardown as well. `_terminate_pipeline_remaining_stages`
has the same cancellation hazard the timeout path had: it fans out over
pre-created tasks, but the `gather` awaiting them was unshielded, so a caller
cancelling mid-grace skipped the SIGKILL escalation. Because it terminates via
wait-tasks rather than calling `_terminate_process` directly, it did not match
the shape the earlier fix covered. Extract the shielding into
`_await_teardown_shielded` and route both paths through it. The new test is
mutation-verified: restoring the bare `gather` fails it with the stage alive.

Adopt the pre-existing shared `_RunKwargs`/`ExecuteFn` aliases from
`tests/helpers/execution`, which `test_safe_cmd_stdin` and
`test_safe_cmd_context` already use. The previous round dismissed the
`dict[str, typ.Any]` at these sites as irreducibly heterogeneous and invented a
local TypedDict; that was wrong on both counts — a shared alias covering
exactly these keywords already existed. This removes the duplicate and the last
`typ.Any` at the timeout-test sites.
